+++ modules/foundry_script/tests/scripts/runtime/features/custom_hash_collision_key.notest.fs
@@ -6,5 +6,5 @@ func _init(p_id: int, p_label: String) -> void:
 	super(p_id, p_label)
 
 
-func _hash_code() -> int:
+func _hash_code() -> long:
 	return 42
